Module 5 - fix project structure
