+++ Module-5/services/common-service.js
@@ -1,0 +1,4 @@
+export const handleMalformedRequest = (res, { code, message }) => {
+  res.statusCode = code;
+  res.end(message);
+};
