+++ Module-5/services/user-service.js
@@ -36,8 +36,3 @@ export const updateUserHobbies = (id, body) => {
   }
   return false;
 };
-
-export const handleMalformedRequest = (res, { code, message }) => {
-  res.statusCode = code;
-  res.end(message);
-};
